reorganize clients
