+++ Cargo.toml
@@ -7,6 +7,9 @@ license-file = "LICENSE"
 repository = "https://github.com/dfinity/ic-bn-lib"
 readme = "README.md"
 
+[features]
+clients-hyper = ["hyper-rustls"]
+
 [dependencies]
 ahash = "0.8.11"
 anyhow = "1.0.93"
@@ -37,7 +40,7 @@ http-body = "1.0.1"
 http-body-util = "0.1.2"
 humantime = "2.2.0"
 hyper = "1.6.0"
-hyper-rustls = { version = "0.27.5", default-features = false, features = [
+hyper-rustls = { version = "0.27.5", optional = true, default-features = false, features = [
     "http1",
     "http2",
     "tls12",
