+++ src/http/client/mod.rs
@@ -1,54 +1,24 @@
 pub mod cli;
+#[cfg(feature = "clients-hyper")]
+pub mod clients_hyper;
+pub mod clients_reqwest;
 
-use std::{
-    fmt,
-    sync::{
-        Arc,
-        atomic::{AtomicUsize, Ordering},
-    },
-    time::{Duration, Instant},
-};
+use std::{fmt, sync::Arc, time::Duration};
 
-use ahash::RandomState;
-use anyhow::{Context, anyhow};
 use async_trait::async_trait;
 use bytes::Bytes;
-use http::{header::HeaderValue, uri::Scheme};
+use http::HeaderValue;
 use http_body_util::Full;
 use hyper::body::Incoming;
-use hyper_rustls::HttpsConnector;
-use hyper_util::{
-    client::legacy::{
-        Client as ClientHyper,
-        connect::{
-            HttpConnector,
-            dns::{GaiResolver, Name as HyperName},
-        },
-    },
-    rt::{TokioExecutor, TokioTimer},
-};
-use moka::sync::{Cache, CacheBuilder};
+use hyper_util::client::legacy::connect::dns::Name as HyperName;
 use prometheus::{
     HistogramVec, IntCounterVec, IntGaugeVec, Registry, register_histogram_vec_with_registry,
     register_int_counter_vec_with_registry, register_int_gauge_vec_with_registry,
 };
 use reqwest::{Request, Response, dns::Resolve};
-use scopeguard::defer;
 use tower::Service;
-use url::Url;
 
-use super::{Error, dns::Resolver};
-
-// Extracts host:port from the URL
-fn extract_host(url: &Url) -> String {
-    format!(
-        "{}:{}",
-        url.host_str()
-            .and_then(|x| x.split('@').next_back())
-            .unwrap_or_default(),
-        url.port_or_known_default().unwrap_or_default()
-    )
-}
+use super::Error;
 
 /// Generic HTTP client trait that is using Reqwest types
 #[cfg_attr(test, mockall::automock)]
@@ -161,354 +131,6 @@ impl<R: CloneableDnsResolver> Default for Options<R> {
     }
 }
 
-pub fn new<R: CloneableDnsResolver>(opts: Options<R>) -> Result<reqwest::Client, Error> {
-    let mut client = reqwest::Client::builder()
-        .connect_timeout(opts.timeout_connect)
-        .read_timeout(opts.timeout_read)
-        .timeout(opts.timeout)
-        .pool_idle_timeout(opts.pool_idle_timeout)
-        .tcp_nodelay(true)
-        .tcp_keepalive(opts.tcp_keepalive)
-        .http2_keep_alive_interval(opts.http2_keepalive)
-        .http2_keep_alive_timeout(opts.http2_keepalive_timeout)
-        .http2_keep_alive_while_idle(opts.http2_keepalive_idle)
-        .http2_adaptive_window(true)
-        .user_agent(opts.user_agent)
-        .redirect(reqwest::redirect::Policy::none())
-        .no_proxy();
-
-    if let Some(v) = opts.pool_idle_max {
-        client = client.pool_max_idle_per_host(v);
-    }
-
-    if let Some(v) = opts.tls_config {
-        client = client.use_preconfigured_tls(v);
-    }
-
-    if let Some(v) = opts.dns_resolver {
-        client = client.dns_resolver(Arc::new(v));
-    }
-
-    Ok(client.build().context("unable to create reqwest client")?)
-}
-
-#[derive(Clone, Debug)]
-pub struct ReqwestClient(reqwest::Client);
-
-impl ReqwestClient {
-    pub fn new<R: CloneableDnsResolver>(opts: Options<R>) -> Result<Self, Error> {
-        Ok(Self(new(opts)?))
-    }
-}
-
-#[async_trait]
-impl Client for ReqwestClient {
-    async fn execute(&self, req: Request) -> Result<Response, reqwest::Error> {
-        self.0.execute(req).await
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct ReqwestClientRoundRobin {
-    inner: Arc<ReqwestClientRoundRobinInner>,
-}
-
-#[derive(Debug)]
-struct ReqwestClientRoundRobinInner {
-    cli: Vec<reqwest::Client>,
-    next: AtomicUsize,
-}
-
-impl ReqwestClientRoundRobin {
-    pub fn new<R: CloneableDnsResolver>(opts: Options<R>, count: usize) -> Result<Self, Error> {
-        let inner = ReqwestClientRoundRobinInner {
-            cli: (0..count)
-                .map(|_| new(opts.clone()))
-                .collect::<Result<Vec<_>, _>>()?,
-            next: AtomicUsize::new(0),
-        };
-
-        Ok(Self {
-            inner: Arc::new(inner),
-        })
-    }
-}
-
-#[async_trait]
-impl Client for ReqwestClientRoundRobin {
-    async fn execute(&self, req: Request) -> Result<Response, reqwest::Error> {
-        let next = self.inner.next.fetch_add(1, Ordering::SeqCst) % self.inner.cli.len();
-        self.inner.cli[next].execute(req).await
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct ReqwestClientLeastLoaded {
-    inner: Arc<Vec<ReqwestClientLeastLoadedInner>>,
-    metrics: Option<Metrics>,
-}
-
-#[derive(Debug)]
-struct ReqwestClientLeastLoadedInner {
-    cli: reqwest::Client,
-    outstanding: Cache<String, Arc<AtomicUsize>, RandomState>,
-}
-
-impl ReqwestClientLeastLoaded {
-    pub fn new<R: CloneableDnsResolver>(
-        opts: Options<R>,
-        count: usize,
-        registry: Option<&Registry>,
-    ) -> Result<Self, Error> {
-        let inner = (0..count)
-            .map(|_| -> Result<_, _> {
-                Ok::<_, Error>(ReqwestClientLeastLoadedInner {
-                    cli: new(opts.clone())?,
-                    // Creates a cache with some sensible max capacity to hold target hosts.
-                    // If the host isn't contacted in 10min then we remove it.
-                    // TODO should we make this configurable? Probably ok like this
-                    outstanding: CacheBuilder::new(16384)
-                        .time_to_idle(Duration::from_secs(600))
-                        .build_with_hasher(RandomState::default()),
-                })
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-
-        Ok(Self {
-            inner: Arc::new(inner),
-            metrics: registry.map(Metrics::new),
-        })
-    }
-}
-
-#[async_trait]
-impl Client for ReqwestClientLeastLoaded {
-    async fn execute(&self, req: Request) -> Result<Response, reqwest::Error> {
-        // Extract host:port from the request URL
-        let host = extract_host(req.url());
-        let labels = &[&host];
-
-        self.metrics
-            .as_ref()
-            .inspect(|x| x.requests.with_label_values(labels).inc());
-
-        // Select the client with least outstanding requests for the given host
-        let (cli, counter) = self
-            .inner
-            .iter()
-            .map(|x| {
-                (
-                    &x.cli,
-                    // Get an atomic counter for the given host or create a new one
-                    x.outstanding
-                        .get_with_by_ref(&host, || Arc::new(AtomicUsize::new(0))),
-                )
-            })
-            .min_by_key(|x| x.1.load(Ordering::SeqCst))
-            .unwrap();
-
-        // The future can be cancelled so we have to use defer to make sure the counter is decreased
-        defer! {
-            counter.fetch_sub(1, Ordering::SeqCst);
-            self.metrics
-                .as_ref()
-                .inspect(|x| x.requests_inflight.with_label_values(labels).dec());
-        }
-
-        counter.fetch_add(1, Ordering::SeqCst);
-        self.metrics
-            .as_ref()
-            .inspect(|x| x.requests_inflight.with_label_values(labels).inc());
-
-        // Execute the request & observe duration
-        let start = Instant::now();
-        let result = cli.execute(req).await;
-        self.metrics.as_ref().inspect(|x| {
-            x.request_duration
-                .with_label_values(labels)
-                .observe(start.elapsed().as_secs_f64())
-        });
-
-        result
-    }
-}
-
-impl Stats for ReqwestClientLeastLoaded {
-    fn stats(&self) -> ClientStats {
-        ClientStats {
-            pool_size: self.inner.len(),
-            outstanding: self
-                .inner
-                .iter()
-                .flat_map(|x| x.outstanding.iter().map(|x| x.1.load(Ordering::SeqCst)))
-                .sum(),
-        }
-    }
-}
-
-impl ClientWithStats for ReqwestClientLeastLoaded {
-    fn to_client(self: Arc<Self>) -> Arc<dyn Client> {
-        self
-    }
-}
-
-#[derive(Debug)]
-pub struct HyperClient {
-    cli: ClientHyper<HttpsConnector<HttpConnector<GaiResolver>>, Full<Bytes>>,
-}
-
-impl HyperClient {
-    pub fn new(opts: Options<Resolver>) -> Self {
-        let mut http_conn = HttpConnector::new_with_resolver(opts.dns_resolver.unwrap());
-        http_conn.set_connect_timeout(Some(opts.timeout_connect));
-        http_conn.set_keepalive(opts.tcp_keepalive);
-        http_conn.enforce_http(false);
-        http_conn.set_nodelay(true);
-
-        let builder = HttpsConnector::<HttpConnector>::builder();
-        let builder = if let Some(mut v) = opts.tls_config {
-            v.alpn_protocols = vec![];
-            builder.with_tls_config(v)
-        } else {
-            builder.with_webpki_roots()
-        };
-
-        let https_conn = builder.https_or_http().enable_all_versions().build();
-
-        let mut builder = ClientHyper::builder(TokioExecutor::new());
-        builder.http2_adaptive_window(true);
-        builder.pool_idle_timeout(opts.pool_idle_timeout);
-
-        if let Some(v) = opts.pool_idle_max {
-            builder.pool_max_idle_per_host(v);
-        }
-
-        let cli = builder
-            .pool_timer(TokioTimer::new())
-            .timer(TokioTimer::new())
-            .build(https_conn);
-
-        Self { cli }
-    }
-}
-
-#[async_trait]
-impl ClientHttp for HyperClient {
-    async fn execute(
-        &self,
-        req: http::Request<Full<Bytes>>,
-    ) -> Result<http::Response<Incoming>, Error> {
-        self.cli
-            .request(req)
-            .await
-            .map_err(|e| Error::Generic(anyhow!("Error executing HTTP request: {e:#}")))
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct HyperClientLeastLoaded {
-    inner: Arc<Vec<HyperClientLeastLoadedInner>>,
-    metrics: Option<Metrics>,
-}
-
-#[derive(Debug)]
-struct HyperClientLeastLoadedInner {
-    cli: HyperClient,
-    outstanding: Cache<String, Arc<AtomicUsize>, RandomState>,
-}
-
-impl HyperClientLeastLoaded {
-    pub fn new(
-        opts: Options<Resolver>,
-        count: usize,
-        registry: Option<&Registry>,
-    ) -> Result<Self, Error> {
-        let inner = (0..count)
-            .map(|_| -> Result<_, _> {
-                Ok::<_, Error>(HyperClientLeastLoadedInner {
-                    cli: HyperClient::new(opts.clone()),
-                    // Creates a cache with some sensible max capacity to hold target hosts.
-                    // If the host isn't contacted in 10min then we remove it.
-                    outstanding: CacheBuilder::new(16384)
-                        .time_to_idle(Duration::from_secs(600))
-                        .build_with_hasher(RandomState::default()),
-                })
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-
-        Ok(Self {
-            inner: Arc::new(inner),
-            metrics: registry.map(Metrics::new),
-        })
-    }
-}
-
-#[async_trait]
-impl ClientHttp for HyperClientLeastLoaded {
-    async fn execute(
-        &self,
-        req: http::Request<Full<Bytes>>,
-    ) -> Result<http::Response<Incoming>, Error> {
-        let uri = req.uri();
-        let host = uri.host().unwrap_or_default();
-        let port = uri.port_u16().unwrap_or_else(|| {
-            if uri.scheme() == Some(&Scheme::HTTPS) {
-                443
-            } else if uri.scheme() == Some(&Scheme::HTTP) {
-                80
-            } else {
-                0
-            }
-        });
-        let host = format!("{host}:{port}");
-
-        let labels = &[&host];
-
-        self.metrics
-            .as_ref()
-            .inspect(|x| x.requests.with_label_values(labels).inc());
-
-        // Select the client with least outstanding requests for the given host
-        let (cli, counter) = self
-            .inner
-            .iter()
-            .map(|x| {
-                (
-                    &x.cli,
-                    // Get an atomic counter for the given host or create a new one
-                    x.outstanding
-                        .get_with_by_ref(&host, || Arc::new(AtomicUsize::new(0))),
-                )
-            })
-            .min_by_key(|x| x.1.load(Ordering::SeqCst))
-            .unwrap();
-
-        // The future can be cancelled so we have to use defer to make sure the counter is decreased
-        defer! {
-            counter.fetch_sub(1, Ordering::SeqCst);
-            self.metrics
-                .as_ref()
-                .inspect(|x| x.requests_inflight.with_label_values(labels).dec());
-        }
-
-        counter.fetch_add(1, Ordering::SeqCst);
-        self.metrics
-            .as_ref()
-            .inspect(|x| x.requests_inflight.with_label_values(labels).inc());
-
-        // Execute the request & observe duration
-        let start = Instant::now();
-        let result = cli.execute(req).await;
-        self.metrics.as_ref().inspect(|x| {
-            x.request_duration
-                .with_label_values(labels)
-                .observe(start.elapsed().as_secs_f64())
-        });
-
-        result
-    }
-}
-
 pub fn basic_auth<U, P>(username: U, password: Option<P>) -> HeaderValue
 where
     U: fmt::Display,
@@ -530,42 +152,4 @@ where
     let mut header = HeaderValue::from_bytes(&buf).expect("base64 is always valid HeaderValue");
     header.set_sensitive(true);
     header
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_extract_host() {
-        assert_eq!(
-            extract_host(&Url::parse("https://foo:123/bar/beef").unwrap()),
-            "foo:123"
-        );
-
-        assert_eq!(
-            extract_host(&Url::parse("https://foo:443/bar/beef").unwrap()),
-            "foo:443"
-        );
-
-        assert_eq!(
-            extract_host(&Url::parse("https://foo/bar/beef").unwrap()),
-            "foo:443"
-        );
-
-        assert_eq!(
-            extract_host(&Url::parse("http://foo:80/bar/beef").unwrap()),
-            "foo:80"
-        );
-
-        assert_eq!(
-            extract_host(&Url::parse("http://foo/bar/beef").unwrap()),
-            "foo:80"
-        );
-
-        assert_eq!(
-            extract_host(&Url::parse("https://top:secret@foo:123/bar/beef").unwrap()),
-            "foo:123"
-        );
-    }
 }
